requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pkg_resources==0.0.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
